feat: add custom hook useCanvasEditor for context access
- Implemented useCanvasEditor hook to provide easy access to CanvasEditorContext.
- Added error handling to ensure the hook is used within the CanvasEditorProvider.

chore: add vitest configuration for testing

- Created vitest.config.js to set up testing environment with jsdom.
