+++ apps/display_editor/package.json
@@ -6,30 +6,36 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
+    "test": "vitest",
     "lint": "eslint . || true",
     "preview": "vite preview",
     "format": "prettier --write ."
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.1020.0",
-    "@tanstack/react-query": "^5.90.3",
     "@repo/assets": "*",
+    "@tanstack/react-query": "^5.90.3",
     "axios": "1.13.6",
     "html-dom-parser": "^7.0.0",
     "i": "^0.3.7",
     "lucide-react": "^0.575.0",
     "npm": "^11.12.1",
-    "react": "^19.2.0",
-    "react-dom": "^19.2.0",
+    "react": "latest",
+    "react-dom": "latest",
     "react-konva": "^19.2.2",
     "use-image": "^1.1.4"
   },
   "devDependencies": {
     "@repo/eslint-config": "*",
+    "@testing-library/react": "^16.3.2",
+    "@testing-library/user-event": "^14.6.1",
     "@types/react": "^19.2.7",
     "@types/react-dom": "^19.2.3",
     "@vitejs/plugin-react": "^5.1.1",
     "babel-plugin-react-compiler": "^1.0.0",
-    "vite": "^7.3.1"
+    "vite": "^7.3.1",
+    "vitest": "^4.1.5",
+    "jsdom": "^29.1.1",
+    "@testing-library/jest-dom": "^6.9.1"
   }
 }
