+++ apps/display_editor/vitest.config.js
@@ -1,0 +1,8 @@
+import { defineConfig } from 'vite';
+
+// https://vite.dev/config/
+export default defineConfig({
+  test: {
+    environment: 'jsdom',
+  },
+});
